Add functionality for command 'movie-this'
The node command 'liri movie-this <movie title>' successfully makes a
call using axios to the OMDB API and pull data.  The data is then logged
to the console in easier readability than JSON and only pulls back
relevant data for the movie.

Next thing to add is pull a default  movie title when the user does not give the
app a movie title.