Find concerts function complete
I have made an API call to the bandsintown API and am pulling data using
the axios node package.

It was difficult to display desired data so I had to use a for loop to
iterate over the objects.  After looping through the data I was able to
select desired information to be displayed.

I used moment.js to format the date.  Now when the user inputs node liri
concert-this <band name>, it returns the name of the band, the name of
the venue, the location, and the date.