Loop over spotify API
I created a for loop to loop over the large amount of data retrieved
when making an API call for a song.

I need to retrieve specific pieces of data and make it look 'pretty' in
the console as this is a CLI application.

I would also like implement colors.js node package in the future.