spotify-this-song command working
Successfully grabbing desired data when the Spotify API is called.  Need
to implement a default search term for this function as well as other
command function.

I also need to use the fs package to write a log everytime a user enters
a command, and the 'do-what-it-says' command which will display the last
command the user inputted to the console.