Add getTweets function using Twitter API
I have decided to add an additional function to my LIRI Bot.  I want the
user to be able to type the command get-tweets <tweetsubject> and be
able to return the top 10 tweets of the subject of their choice.

I have built the function and included the twitter node package in package.json, when I try and run the function
I get 'invalid or expired  api token' error message.  I will sign up for
a new API key and retry the function.