Returning tweet details
I am returning the desired details by looping over data.statuses and
then creating variables for the details I want to display in the
console.  Some results are coming up undefined.  This is either because
I didn't correctly call the variable or some of the results don't have
some results to display.  I will look into this for a future update.