Implement colors package
I have created a theme to turn the color of the text for this
application green, and red for errors.  I had to use the colors npm
string safe API in order to be able to just put the desired color in
front of a console.log.

I need to reformat my multiple console.logs in my functions to one large
console.log to avoid having to repeatedly add a color to the message.
It should look similar to the function movieThis console.log when I am
finished.

I also need to incorporate the fs package for the liri do-what-it-says
command before this application will be complete.