do-this function is working
I got the command to read the random.txt file, then turn the data into
an array, split by commas.  After that, I call the startApp function with
the paramters of index[0] for the command and index[1] for the input.

I changed the random.txt file to another command in order to test that
it still works.

I want to attempt to log every command the user inputs to a log.txt
file.  I can use fs.appendFile to achieve these ends.  I would also like
to add a few more UI enhancements for easier readability.  Last thing I
need to do is add a GIF of the app in action to add to the README.