Replace colors package with chalk
I decided to ditch the colors package with a more popular chalk package
for styling.  It has an option to output a bright green color which
looks similar to The Matrix theme which I am shooting for with this CLI
application.

Next I would like to create themese like I did with the colors package
to save typing out the entire color name of greenBright.

Since this application has many console.log methods, I saved that method
to a variable called log for cleaner code.